Change priority and destination lists to sets
By using sets instead of lists for the valid priority and destination,
the checks for item uniqueness in these datatypes is now handled
inherently, with little detriment.

The only drawback to this is that the system that chooses the default
priority for each tool now simply chooses the first element from
iter(), rather than the one in the middle of the list. This isn't a
huge deal, as that was no more than a guess at what a mid-level
priority would be, and would even cause errors when the chosen
priority wasn't in the tool's default destination section.

Using sets insetad still result in the same issue, so this will need
to be fixed soon, as it prevents certain tool destination configs
from working properly.