Make users adhere to same priorities as tools, better handle tools without default priority destination, stop destination and priority sets from retaining values between test cases, updating tests
Previously, user's had their own sets of priorities (low, med, high)
that did not match the new system with configurable priorities. As
mapping these user priorities onto tool priorities would be impossible
without extra config sections, or hardcoding of things that map to one
another, the user priorities are now on the same system. IE: all
priorities declared in the user sections much match priorities in the
"default_destination" section.

Dynamic_tool_destination also now uses a less-hacky method of handling
tools that don't have a default destination for the default priority
(which is still unfortunately arbitrarily chosen; this presents a
rather large issue that will need to be addressed in the future.)
Essentially, the global default destination for the default priority
is used if there isn't one defined in the tool's defaults or its rules.

There was also a bug previously that caused dynamic_tool_destination's
priority_list and and destination_list (formerly valid_destinations)
to retain values in between test cases, causes the test results to
differ from the expected results. This bug has been fixed.

Finally, the test_dynamic_tool_destination program has been updated
to reflect the exptected outcome of test cases with the new program's
expected functionality in mind. Note that this does not mean the tests
are testing what they should be -- they simply don't fail anymore for
expected behaviour.
There is test that still fails -- and it deals with the fact that
there no longer is a default priority that all tools need to adhere
to. Because default priorities are determined now in an arbitrary way
(this shouldn't be permanent behaviour), a destination other than the
expected one may be chosen for a given tool that does not implement
that (arbitrarily chosen) default priority.