Alter error messages, add new tests
A few error messages were altered to start with lowercase letters.
Why? Don't ask me, that was yesterday. Expect all the error
messages to be altered to start with capitals (as I believe they
should) soon.

Two new tests were added, ensuring that errors are fired when
priorities are referenced that don't appear in the config's global
default_destinaion section.