Allow job config xml files to be passed as an argument
job configuration xml files can now be passed as arguments to the
application with the -j  or the --job-config options. To
facilitate this change, the parse_yaml constructor has been altered
to accept a new parameter, job_conf_path, the path to the job
configuration.

For simplicity, get_destination_list_from_job_config is now called
from parse yaml instead of validate_config. This simply reduces the
amount of times the job_conf path needs to be passed between
functions.