Add ability to set default priority, update tests, small fixes
This is a relatively large commit -- a new section in
tool_destinations.yml is now required: default_priority, which sets
which priority to be used as default for all tools, replacing the
way 'med' was previously assumed to be default, back when
priorities were not configurable.

As a fallback to ensure compatibility with the old style of config,
if no default_priority is specified in the config, but 'med' exists
as a valid priority, the priority will be set to 'med'.

Additionally, tests were updated to expect this new behaviour.

A collection of small bug fixes were also added, such as:
 - not adding anything to the list of priorities when the global
   default_destination is only a string,
 - Making error messages clearer
 - fixing some indentation