Add new tests, add new category to valid keys check
Added three new tests, testing cases where all destinations are str,
where the default_destination is a str but a tool
default_destination is a dict, and where the default_destination is
a str but a tool rule destination is a dict.

Also fixed a bug that was caught with these new tests --
default_priority was not actually in the list of valid config
categories so using the default_priority key would result in errors.
Whoops.