Allow per-request SQL debugging via URL parameter.
This has been super helpful in debugging the SQL generated by Galaxy for recent PRs.

To use this simply enable the middleware, use your browser's network logger to find the URL of interest, copy the URL into the browser address bar and add ``?debug_sql=1`` to it. You will then see the sql in the logs only for that request and for every statement in that request.

Simply turning all logging on results in a deluge of logging I've never been able to grok and that is very hard to ping to particular request. Slow request logging is great but it doesn't make this kind of per-query full profile of the SQL generated as easy as this approach and isn't great at debugging performance problems when the result from a large number of small requests (you'd have to set a very low threshold to make sure you are logging every query and then it is no different from the full log).