Begin adding system to suggest typo corrections
Implemented a function to check for edit distance that I found
online, and wrote a function that compares a given string with all
entries in a set and returns the one with fewest edits.

Added code to validate_destination in RuleValidator to suggest
corrections for possibly mispelled priorities, but haven't tested
it yet.