Finish implementing typo correction suggestions
Code has been added in all locations in the code that check values
against the list of valid values, and checks if there are similar
valid values in the list if the value in the config is invalid.

If there is a similar value (edit distance 3 or closer, currently,)
output will be printed pointing out the similar, valid value.