Better handling of no job configs, fixing indentation
Now, if no job config is specified and none can be found in the
'config/' folder, rather than crashing trying to access a file that
doesn't exist, the program just keeps the destination_list empty,
resulting in a bunch of "destination not found" errors (as expected,)
but (hopefully) no crashing.

Also fixed weird 2-space indentation that happened in the previous
few commits.