Create method for verifying destinations
Now, destinations are validated in a function that returns
both the validity of destination, as well as a suggestion
for typo correction if a typo is suspected. This function
is what does the check for if `app` exists now, making
validating destinations use much less lines of code now,
and also making these sections of code much easier to read.