Add a condition to FilterFailedDatasetsTool to remove empty files
Hello,

Here is an extra-condition for the tool ``FilterFailedDatasetsTool`` to allow it to remove empty datasets as well as failed datasets in a collection. I tested it on my data and it seems to work fine. Since it's my first time looking at galaxy's source code, I don't know if there is a proper way to do that ... And I didn't make the improvement for paired datasets because I don't have proper data to test it. But I assume it would be something like ``if forward_o.is_ok and reverse_o.is_ok and forward_o.has_data() and reverse_o.has_data():`` ?