Add New Tests, Fix Bug Uncovered From Test
Added two new tests to test for `valid_config` not being messed with by
previous assignments to it.

This proved that the previous design of setting `valid_config` to the
value of `validate_destination` in `validate_config` caused errant
behaviour when a default destination was invalid but a tool default
destination was valid. This case caused the program to still end up
returning `valid_config` as True even when an invalid default destination
was found.