Use target file path when adding missing indexes
Previously we would load a TS tool during installation and ask if any of its
tool datat tables are missing loc files. If thats the case we copy the sample
file and call `handle_found_index_file` for the particular missing *.loc path
of this data table.

This does not work if the data table already existed (e.g for built-in data
tables), but was missing the *.loc file, since we would atempt to load that
path. Instead we do now load the path to the loc file that we just copied.

Should fix https://github.com/galaxyproject/galaxy/issues/5648.