Fixes for testing Dockerfile recipe.
The current Dockerfile build is broken because Selenium is not compatible with the latest chromedriver it seems (and the version in galaxy/testing-base:18.01.4 isn't compatible with the latest stable Chrome - version 65). I can't get older versions of Chrome to install with the current apt source but this does at least pin the version of chromedriver.

Longer term it would be good to have the version of chrome we target to be pinned also, I hate that this Dockerfile isn't reproducible.