Make problematic test test_library_to_collections more robust potentially.
Now that #5701 has been merged we have some more clues on the problem with this test. It is dying on the second upload being changed here. I suspect what is happening is that we are waiting on a history to become okay but we make that check before the first upload has even finished - so there is some chance the history is okay because it is empty - so we are reopening the home page in the middle of the first upload and this is causing something odd to happen - something is popping up an empty alert. I'm not sure what is causing that alert but it isn't what we are testing with this test so it is best to really just wait on the uploads - as we do in other tests (see test_uploads.py for instance).

history_panel_wait_for_hid_ok is a stronger assertion and is generally more robust and has better error reporting anyway so this should be a stability win all around.