[18.01] Fix SQL query for "workflows shared with me" as part of workflow index.
Before and after queries here:

https://gist.github.com/jmchilton/dc57bca60178341d9ce0f6cf6d4bd36f

You can see that before the joined load of the stored workflow wasn't being registered by SQLalchemy properly and so it was joining that table twice - once using the proper index and another time just walking the whole table. The filter and order by were being applied to the unconditional version - kind of like saying give me all the workflows and the right workflows and I will check all the workflows to see if this one is deleted instead of just the subset of workflows that are actually being shared with the target user.