Skip dataset conversion validation at TS install
During TS installation tools will be loaded using ValidationContext
that only has the bare minimum required for loading tools. Notably
it does not have acces to the loaded datatypes, and therefore we can't
validate the conversion tag used for example in freebayes.
If we're dealing with a ValidationContext we now ignore missing
conversion targets.