Styles: Improving colors and clearning up config.
Hardcoded the old derived colors to use as a base, and then deriving
from those. We may want to update even the base colors (blue, green,
etc) using thise. Alert backgrounds are now very close. Since these
are the history background colors they are the most important to match
to maintain the look. Removed lots of configuration variables that are
not used anymore for bootstrap 4.