Change 'a when tag has not been defined for' level to debug
We've discussed this for a while going back and forth over this.
Ultimately it's probably not that useful for admins and it clutters
the warnings. There is a way to configure logging in planemo
based on the python module (https://github.com/galaxyproject/planemo/commit/1eaff26f7c2db1e6eff25b98abbad19e4dc2be56)
and planemo lint will catch this anyway.