Optimize fetching collections for tool form building.
Low hanging fruit is to exclude hidden collections, that probably won't buy much for typical uses.

This eliminates one extra tag query per dataset collection that appears in the rendered result, this probably buys us a bit more than the hidden collection thing but is still probably a good choice (unless there collections with a large number of tags :(...).

This also eliminates the extra fetch of the first, outer-est collection associated the history dataset collection - not its elements just the collection. This saves a number of queries roughly equal to the number of HDCAs in the history and unlike the tag thing there is no downside really here - there will always be one collection.

Before and after profiling of a tool form build:

https://gist.github.com/jmchilton/d68565662f7f4b7ee2640f09fbb92962