Pause jobs with invalid inputs
Otherwise these jobs may hang around forever/until the admin
does a cleanup. This is probably preferable for users as well,
as the pause message gives a hint of why their job didn't start
running. Fixes https://github.com/galaxyproject/galaxy/issues/6022
and would minimize maintenance as outlined in https://github.com/galaxyproject/galaxy/issues/5944.

I'm working on tests, but one easy way to test this is create a collection
through the list builder interface, then delete the source HDAs.
Then run a workflow on the collection and without this commit
the outputs get created but the job just hangs around indefinetely.
With the PR the steps consuming the collection and the subsequent jobs
will be paused.

Set to paused