[18.05] Fix matches_any() checks for dynamic compressed types.
Check for matches by making sure compressed datatypes match and then using matches_any on the uncompressed datatype. This way for instance fastqsanger.gz matches_any fastq.gz, but doesn't matches_any a fastqsanger. This should fix implicit conversion and such.

I don't feel great about this, this is first instance of datatypes overriding the default matches_any() behavior but on balance I think it is the right thing to do. The alternative I think is to give up on dynamic classes for compressed types and stick with completely custom type heirarchies on per compressed class basis. That process requires some pretty tricky Python type hierarchies (for instance the genbank.gz PR I believe was likely broken in the same way as this was because the gz datatype descended from the genbank type even though it wasn't a proper is-a relationship and matches_any wasn't overridden to reflect this).

This is really getting in the weeds here but other ideas I had included to "build" the compressed class by taking the parent class of the target datatype and re-adding all the matching methods from the target datatype - this would have fixed some false positive matches_any behavior (e.g. fastqsanger.gz matching a fastqsanger) but it would introduced false negatives (e.g. fastqsanger.gz not matching a fastq.gz) or doing something similar with annotations on the actual classes (but that would have a similar problem). One could imagine fixing that by building all compressed classes after all the uncompressed classes are built and walking the whole datatype hierarchy and trying to pick up leafs vs. interior nodes and recreating the hierarchy that way. That would be much more complicated and would lean into some of my least favorite aspects of Galaxy datatypes being so heavily dictated by Python classes and not declarative structures like the XML that can be externally reasoned about.