Fix append() for ANY_COLLECTION_TYPE type description in WF editor.
Append means take a collection type and extend it by the current type - so appending to list to paired -> list:paired for instance. ANY_COLLECTION_TYPE is a representation of incomplete information - it means this type description might be any kind of collection type (the way we use "input" to mean any kind of datatype in the editor).

This is really a second order approximation of what a correct solution would be - a first order approximation would somehow represent the result of append being at least of "depth" specified (list:* or *:list instead of * for instance). A not approximation would be to do that but use ANY_COLLECTION_TYPE much less frequently by capturing type_source information and propagating it throughout the workflow properly. But these are harder things to do, this incorrect fix will allow many valid workflows to be editted that currently cannot be so I think it is good.

"Fixes" #5205.