Use separate `pid_log_paster_args` variable
to store --pid-file and --log-file options of `./scripts/paster.py` .

https://github.com/galaxyproject/galaxy/pull/6239 fixed 3 bugs:

- `GALAXY_RUN_ALL=1 ./run.sh` was reusing the same pid and log files for
  all Galaxy processes;
- `./run.sh restart` was always using `paster.pid` and `paster.log` instead
  of the configured files
- `./run.sh status` was always using `paster.pid`

but also introduced a regression, i.e. the plain `./run.sh` started
writing to the log file instead of the standard output (console).

This fixes the regression by using a separate variable for these options
instead of conflating them in `paster_args` or `server_args`.