Allow invalid optional param values for old tools
For tools with profile < 18.09 we return None for provided, but illegal
default values. This can happen when referencing columns of an optional
dataset for instance, as experienced in https://github.com/galaxyproject/tools-iuc/pull/1842.
This used to pass prior to 18.05 because we would always return the
value if there were no legal values. This was changed in
https://github.com/galaxyproject/galaxy/commit/9febc519944dcfb5c8907b8a302fd6aff6f0f79f
where we would only allow this in workflow building mode.