Reorder inputs before multiplications
This should be a stable order. Ideally this would be configurable to the
user or tool auther, but for now this is based on the order within which
data paramaters appear in tool.inputs, which depends on the order in the
XML/YML.

If this isn't stable the order of the input combinations may change, and
elements will jump in the collection structure. For instance the
reverse:forward output for a pair:pair in inputA multiplied by pair:pair
in input B may end up in forward:reverse if re-run.