Make discovery of cachable jobs more robust
We update the actually used version once the job inputs are ready. This
allows finding jobs when a dataset is being new / being updated while
the job is new.  This is also more accurate since we record the dataset
versino that is actually being used when the tool is being evaluated.

This should allow chaching when simultaneously running workflows that
share steps.