Added support for setting the snpeff_version of a snpeffdb manually uploaded to a user's history.
Prior to this change, the snpeff_version was always set to SnpEff4.0 because the only files the version check was made on were those in the data_dir that houses galaxy's snpeff databases. Thus, if a user uploaded a 4.3 .bin file, the version would be defaulted to 4.0 and not show up in the history dropdown of the snpeff eff tool.

Files checked in: lib/galaxy/datatypes/text.py