Add integration test for uploading datasets of known datatype
We upload everything in lib/galaxy/datatypes/test and make sure
that we get the correct datatype.
For datatypes that have a sniffer we use `auto`, otherwise we set the
correct datatype.

This is fairly basic for now, but it would be trivial to extend this
to also verify file integrity and metadata elements as well as testing
options that affect uploads, like `auto_decompress` settings.