Allow ``remove_value`` filter on data tables
The `RemoveValueFilter.filter_options()` method was expecting `options` to
be a list of lists, where each inner list contains exactly 3-elements
(name, value, selected), which is the format returned by the `data_meta`
filter. In fact, the only known use of the `remove_falue` filter is in the
tool `tools/maf/maf_to_interval.xml`, to filter dataset metadata.

This commit generalises the filter to look for a column of the dynamic
options named `value`, and use it during the filtering.

Fix https://github.com/galaxyproject/tools-iuc/issues/2151