datatype_fastg
FASTG is a format for faithfully representing genome assemblies in the
face of allelic polymorphism and assembly uncertainty.
http://fastg.sourceforge.net/FASTG_Spec_v1.00.pdf
This format can be an output for metagenomics applications metapades
and megahit.  It is a required input for Graph2Pro-var application that
generates putative protein sequences for the variant assemblies
described by the FASTG format file.