Fix set_peek method for copied datasets
dataset_assoc is the job to ouptut dataset association that references
the HDA created by a job. If the HDA is copied the copies need to be
updated as well, so here we need to call `set_peek` on dataset, not
dataset_assoc.dataset.

Broken in f4343d876f13e55e92ca228582b4430dcd01e4aa

Thanks for tracking this down @jmchilton!