Do not re-compute job input_ext/dbkey (if unneeded).
Remote metadata computation needs input_ext in the job parameters (like dbkey) so no need to re-loop through the inp_data to recompute it.

Additionally the loop here in discover_outputs in the job or job handler might be running on a different Python version or different Galaxy version than the tool action in the web controller - and this is a dict not a list. Therefore the loop might result in a different value for input_ext. For this reason, it is better to just compute it once and reuse it anyway for a greater guarantee of consistency.