Replace deprecated singularity import with build
And upgrade singularity build image to 2.4.6 (latest on biocontainers).
Overall this is necessary because singularity import 2.3X doesn't work
behind proxies. The change from `Singularity` to `Singularity.def` is
necessary because the script uses the `file` command if the Singularity
file doesn't end in .def or is not exactly Singularity (it is
`/import/Singularity` in our case). `file` isn't available in the
biocontainers image.