Apply metadata setter for default to conditionals
The existing code only expects the case like:

```
<actions>
                <action name="column_names" type="metadata" default="Geneid,${alignment.element_identifier}" />
</actions>
```

whereas a conditional can be inserted within `<actions>` yielding a case like:

```
        <actions>
            <conditional name="format">
                <when value="something_good">
                    <action type="metadata" name="column_names" default="one,two,three" />
                </when>
            </conditional>
        </actions>
```

The fact that this is not handled causes the display logic to fail when `column_names` is set for a tabular dataset inside such a conditional. This patch provides a solution for that case.