Properly deal with missing tools in subworkflows
If a subworkflow is missing tools the entire parent workflow couldn't
be loaded into the editor, because we raised an exception. This is now
only logged as an error. On the user side we now prevent running and
saving such workflows by actually implementing the `get_errors` method
for the SubWorkflowModule class.

This should fix https://github.com/galaxyproject/galaxy/issues/7140.