Do not remove tools from other installed revisions when uninstalling TS repo
Multiple installed revisions of a repository can provide the same version of
a tool, e.g. for the repository

https://toolshed.g2.bx.psu.edu/view/nilesh/rseqc

only one tool out of 22 has a different version in revision 57 with
respect to revision 56.

In this case, when uninstalling one repository, the tools with the same guid
were removed also from the other repository revisions in the
`shed_tool_conf` file. This can be prevented by checking also the tool path.